backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8
# Alternative: pdfplumber==0.10.3

# AI Integration
anthropic==0.7.7

# CORS is included with FastAPI

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
httpx==0.25.2

# Image Processing
Pillow==10.1.0

